# packages/shared-proto/openapi.yaml
openapi: 3.0.3
info:
  title: ACE API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /student/enrollments:
    get:
      operationId: studentListEnrollments
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Student enrollments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentListResponse'
    post:
      operationId: studentEnroll
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollRequest'
      responses:
        '200':
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: ace_access
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema:
        type: string
      description: Double-submit CSRF token. Must match the ace_csrf cookie.
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    EnrollmentListResponse:
      type: object
      required: [examPackageIds]
      properties:
        examPackageIds:
          type: array
          items:
            type: string
    EnrollRequest:
      type: object
      required: [examPackageId]
      properties:
        examPackageId:
          type: string
...